requirements.txt
# Foodle backend — Phase 1 (pin in production via pip-tools or similar)
Django>=4.2,<5
django-tenants>=3.6
djangorestframework>=3.14
django-environ>=0.11
django-cors-headers>=4.3
psycopg[binary]>=3.1
uvicorn[standard]>=0.23
